requirements.txt
allure-pytest==2.14.2
allure-python-commons==2.14.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.3.0
execnet==2.1.1
Faker==37.3.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
playwright==1.52.0
pluggy==1.6.0
psycopg2==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
pytest==8.3.5
pytest-base-url==2.1.0
pytest-check==2.5.3
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.1
pytest-playwright==0.7.0
pytest-rerunfailures==15.1
pytest-xdist==3.6.1
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
text-unidecode==1.3
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
